[initial] Unless I'm wrong, we can condense these two files into one.
As far as I know, without actually testing, getSubscribedEvents actually tells the events
we are subscribed to. Because it's an array, that gives me the idea that there can be more
than one. So, that makes me think the file name is not so important, so we can probably
merge the two into one. Neh?