A couple of changes, neither of which is necessarily complete.
First of all, I removed the raw style field for prefixes. Instead, I plan
to support BBCode for prefixes. Not only is that easier for me to handle,
it is also more consistent and familiar for the average admin.

Second, I changed how tokens are handled and made it easy for other
extensions to add tokens as desired.